% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/makeset.R
\name{makeset}
\alias{makeset}
\title{makeset}
\usage{
makeset(nbgo, nbgenes, nbgopergenes)
}
\arguments{
\item{nbgo}{integer - total number of unique go}

\item{nbgenes}{integer - total number of unique genes}

\item{nbgopergenes}{float - average nb of go that a genes belong to}
}
\description{
Makeset- Create a SIMULATED set of go / genes
}
\details{
This function creates a random bipartite graph simulated gene sets
}
